Airflow_with_Spark/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pyspark>=3.5.0
boto3>=1.28.0
apache-airflow-providers-apache-spark>=4.0.0
apache-airflow-providers-amazon >=9.18.1